refactor(mentor): 커스텀 훅 사용, 폴더에서 폴더 목록과 전체 링크 useEffect를 컴포넌트와 분리
