+++ README.md
@@ -47,22 +47,22 @@
 - [x] 상수는 컴포넌트 외부에서 관리한다.
 
 ## 함수, 메서드
-- [ ] 절차가 단순한 코드를 함수화하는 것은 지양한다.
-- [ ] 한 함수가 한 가지 기능만 담당하게 한다.
-  - [ ] 함수가 한 가지 기능을 하는지 확인하는 기준을 세운다.
-  - [ ] 특히, 함수 역할의 일관성을 고민한다.
-- [ ] 함수형 프로그래밍을 고민해본다. 함수가 외부 변수에 의해 동작하지 않는지 점검한다.
+- [x] 절차가 단순한 코드를 함수화하는 것은 지양한다.
+- [x] 한 함수가 한 가지 기능만 담당하게 한다.
+  - [x] 함수가 한 가지 기능을 하는지 확인하는 기준을 세운다.
+  - [x] 특히, 함수 역할의 일관성을 고민한다.
+- [x] 함수형 프로그래밍을 고민해본다. 함수가 외부 변수에 의해 동작하지 않는지 점검한다.
   - [x] 특히, 전역 변수에 의존하는 이벤트를 점검한다. 이를 지역 변수와 인자로 처리한다.
-  - [ ] 함수의 인자를 최소화한다.
+  - [x] 함수의 인자를 최소화한다.
 ### 네이밍 관련
-- [ ] 함수의 이름과 역할을 분명히 한다.
+- [x] 함수의 이름과 역할을 분명히 한다.
 ### 컨벤션 관련
-- [ ] indent(인덴트, 들여쓰기) depth를 3이 넘지 않도록 구현한다. 2까지만 허용한다. 예를 들어 while문 안에 if문이 있으면 들여쓰기는 2이다. 
-  - [ ] 힌트: indent(인덴트, 들여쓰기) depth를 줄이는 좋은 방법은 함수(또는 메서드)를 분리하면 된다.
-- [ ] 함수(또는 메서드)의 길이가 15라인을 넘어가지 않도록 구현한다.
-- [ ] else를 지양한다.
-    - [ ] 힌트: if 조건절에서 값을 return하는 방식으로 구현하면 else를 사용하지 않아도 된다.
-    - [ ] 때로는 if/else, switch문을 사용하는 것이 더 깔끔해 보일 수 있다. 어느 경우에 쓰는 것이 적절할지 스스로 고민해 본다.
+- [x] indent(인덴트, 들여쓰기) depth를 3이 넘지 않도록 구현한다. 2까지만 허용한다. 예를 들어 while문 안에 if문이 있으면 들여쓰기는 2이다. 
+  - [x] 힌트: indent(인덴트, 들여쓰기) depth를 줄이는 좋은 방법은 함수(또는 메서드)를 분리하면 된다.
+- [x] 함수(또는 메서드)의 길이가 15라인을 넘어가지 않도록 구현한다.
+- [x] else를 지양한다.
+    - [x] 힌트: if 조건절에서 값을 return하는 방식으로 구현하면 else를 사용하지 않아도 된다.
+    - [x] 때로는 if/else, switch문을 사용하는 것이 더 깔끔해 보일 수 있다. 어느 경우에 쓰는 것이 적절할지 스스로 고민해 본다.
 - [ ] 값을 하드 코딩하지 않는다.
 
 ## js
